Move B2CTrustedHostList to B2CAuthority class
