--- lib/msal-common/src/auth/authority/B2CTrustedHostList.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-/**
- * Initializes the B2C Trusted Host List. Populated by AuthorityFactory.setKnownAuthorities(). Used by the B2C authority class to validate authorities.
- */
-export const B2CTrustedHostList: string[] = [];
